+++ processor/src/errors.rs
@@ -1,5 +1,4 @@
-use alloc::{boxed::Box, sync::Arc, vec::Vec};
-use core::error::Error;
+use alloc::{string::ToString, sync::Arc, vec::Vec};
 
 use miden_air::RowIndex;
 use miette::Diagnostic;
@@ -11,27 +10,17 @@ use vm_core::{
 };
 use winter_prover::ProverError;
 
-use super::{
-    Felt, QuadFelt, Word,
-    crypto::MerkleError,
+use crate::{
+    AdviceProviderError, Felt, MemoryError, QuadFelt, Word,
+    handlers::EventError,
     system::{FMP_MAX, FMP_MIN},
 };
-use crate::MemoryError;
 
 // EXECUTION ERROR
 // ================================================================================================
 
 #[derive(Debug, thiserror::Error, Diagnostic)]
 pub enum ExecutionError {
-    #[error("value for key {} not present in the advice map", key.to_hex())]
-    #[diagnostic()]
-    AdviceMapKeyNotFound {
-        #[label]
-        label: SourceSpan,
-        #[source_code]
-        source_file: Option<Arc<SourceFile>>,
-        key: Word,
-    },
     #[error("value for key {} already present in the advice map when loading MAST forest", key.to_hex())]
     #[diagnostic(help(
         "previous values at key were '{prev_values:?}'. Operation would have replaced them with '{new_values:?}'",
@@ -41,14 +30,19 @@ pub enum ExecutionError {
         prev_values: Vec<Felt>,
         new_values: Vec<Felt>,
     },
-    #[error("advice stack read failed at clock cycle {row}")]
+    #[error(
+        "advice provider error at clock cycle {}",
+        .clk.map_or_else(|| "unknown".to_string(), |n| n.to_string()))
+    ]
     #[diagnostic()]
-    AdviceStackReadFailed {
+    AdviceProviderError {
         #[label]
         label: SourceSpan,
         #[source_code]
         source_file: Option<Arc<SourceFile>>,
-        row: RowIndex,
+        #[source]
+        err: AdviceProviderError,
+        clk: Option<RowIndex>,
     },
     /// This error is caught by the assembler, so we don't need diagnostics here.
     #[error("illegal use of instruction {0} while inside a syscall")]
@@ -71,9 +65,9 @@ pub enum ExecutionError {
         source_file: Option<Arc<SourceFile>>,
         clk: RowIndex,
     },
-    #[error("failed to execute the dynamic code block provided by the stack with root 0x{hex}; the block could not be found",
-      hex = to_hex(.digest.as_bytes())
-    )]
+    #[error(
+        "failed to execute the dynamic code block provided by the stack with root {}; the block could not be found",
+        .digest.to_hex())]
     #[diagnostic()]
     DynamicNodeNotFound {
         #[label]
@@ -89,11 +83,12 @@ pub enum ExecutionError {
         label: SourceSpan,
         #[source_code]
         source_file: Option<Arc<SourceFile>>,
+        event_id: u32,
         #[source]
-        error: Box<dyn Error + Send + Sync + 'static>,
+        error: EventError,
     },
     #[error("failed to execute Ext2Intt operation: {0}")]
-    Ext2InttError(Ext2InttError),
+    Ext2InttError(#[from] Ext2InttError),
     #[error("assertion failed at clock cycle {clk} with error {}",
       match err_msg {
         Some(msg) => format!("message: {msg}"),
@@ -131,27 +126,6 @@ pub enum ExecutionError {
         source_file: Option<Arc<SourceFile>>,
         depth: usize,
     },
-    #[error(
-        "provided merkle tree {depth} is out of bounds and cannot be represented as an unsigned 8-bit integer"
-    )]
-    #[diagnostic()]
-    InvalidMerkleTreeDepth {
-        #[label]
-        label: SourceSpan,
-        #[source_code]
-        source_file: Option<Arc<SourceFile>>,
-        depth: Felt,
-    },
-    #[error("provided node index {index} is out of bounds for a merkle tree node at depth {depth}")]
-    #[diagnostic()]
-    InvalidMerkleTreeNodeIndex {
-        #[label]
-        label: SourceSpan,
-        #[source_code]
-        source_file: Option<Arc<SourceFile>>,
-        depth: Felt,
-        index: Felt,
-    },
     #[error("attempted to calculate integer logarithm with zero argument at clock cycle {clk}")]
     #[diagnostic()]
     LogArgumentZero {
@@ -184,7 +158,7 @@ pub enum ExecutionError {
     MastNodeNotFoundInForest { node_id: MastNodeId },
     #[error(transparent)]
     #[diagnostic(transparent)]
-    MemoryError(MemoryError),
+    MemoryError(#[from] MemoryError),
     #[error("no MAST forest contains the procedure with root digest {root_digest}")]
     NoMastForestWithProcedure {
         #[label]
@@ -211,33 +185,6 @@ pub enum ExecutionError {
         root: Word,
         err_code: Felt,
         err_msg: Option<Arc<str>>,
-    },
-    #[error("failed to lookup value in Merkle store")]
-    MerkleStoreLookupFailed {
-        #[label]
-        label: SourceSpan,
-        #[source_code]
-        source_file: Option<Arc<SourceFile>>,
-        #[source]
-        err: MerkleError,
-    },
-    #[error("advice provider Merkle store backend merge failed")]
-    MerkleStoreMergeFailed {
-        #[label]
-        label: SourceSpan,
-        #[source_code]
-        source_file: Option<Arc<SourceFile>>,
-        #[source]
-        err: MerkleError,
-    },
-    #[error("advice provider Merkle store backend update failed")]
-    MerkleStoreUpdateFailed {
-        #[label]
-        label: SourceSpan,
-        #[source_code]
-        source_file: Option<Arc<SourceFile>>,
-        #[source]
-        err: MerkleError,
     },
     #[error("if statement expected a binary value on top of the stack, but got {value}")]
     #[diagnostic()]
@@ -294,7 +241,7 @@ pub enum ExecutionError {
     #[error("a program has already been executed in this process")]
     ProgramAlreadyExecuted,
     #[error("proof generation failed")]
-    ProverError(#[source] ProverError),
+    ProverError(#[from] ProverError),
     #[error("smt node {node_hex} not found", node_hex = to_hex(node.as_bytes()))]
     SmtNodeNotFound {
         #[label]
@@ -333,29 +280,35 @@ pub enum ExecutionError {
         source_file: Option<Arc<SourceFile>>,
         error: AceError,
     },
-}
-
-impl From<Ext2InttError> for ExecutionError {
-    fn from(value: Ext2InttError) -> Self {
-        Self::Ext2InttError(value)
-    }
+    #[error("attempted to add event handler with previously inserted id: {id}")]
+    DuplicateEventHandler { id: u32 },
+    #[error("got unexpected event_id: {id}")]
+    #[diagnostic()]
+    InvalidEventId {
+        #[label]
+        label: SourceSpan,
+        #[source_code]
+        source_file: Option<Arc<SourceFile>>,
+        id: u32,
+    },
 }
 
 impl ExecutionError {
-    pub fn advice_map_key_not_found(
-        key: Word,
+    pub fn advice_error(
+        err: AdviceProviderError,
         err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Self {
         let (label, source_file) = err_ctx.label_and_source_file();
-        Self::AdviceMapKeyNotFound { label, source_file, key }
+        Self::AdviceProviderError { label, source_file, err, clk: None }
     }
 
-    pub fn advice_stack_read_failed(
-        row: RowIndex,
+    pub fn advice_error_at_clk(
+        err: AdviceProviderError,
+        clk: RowIndex,
         err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Self {
         let (label, source_file) = err_ctx.label_and_source_file();
-        Self::AdviceStackReadFailed { label, source_file, row }
+        Self::AdviceProviderError { label, source_file, err, clk: Some(clk) }
     }
 
     pub fn divide_by_zero(clk: RowIndex, err_ctx: &ErrorContext<'_, impl MastNodeExt>) -> Self {
@@ -382,12 +335,19 @@ impl ExecutionError {
     }
 
     pub fn event_error(
-        error: Box<dyn Error + Send + Sync + 'static>,
+        event_id: u32,
+        error: EventError,
         err_ctx: &ErrorContext<'_, impl MastNodeExt>,
     ) -> Self {
         let (label, source_file) = err_ctx.label_and_source_file();
 
-        Self::EventError { label, source_file, error }
+        Self::EventError { label, source_file, event_id, error }
+    }
+
+    pub fn invalid_event_id_error(id: u32, err_ctx: &ErrorContext<'_, impl MastNodeExt>) -> Self {
+        let (label, source_file) = err_ctx.label_and_source_file();
+
+        Self::InvalidEventId { label, source_file, id }
     }
 
     pub fn failed_assertion(
@@ -405,23 +365,6 @@ impl ExecutionError {
             err_code,
             err_msg,
         }
-    }
-
-    pub fn invalid_merkle_tree_depth(
-        depth: Felt,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Self {
-        let (label, source_file) = err_ctx.label_and_source_file();
-        Self::InvalidMerkleTreeDepth { label, source_file, depth }
-    }
-
-    pub fn invalid_merkle_tree_node_index(
-        depth: Felt,
-        index: Felt,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Self {
-        let (label, source_file) = err_ctx.label_and_source_file();
-        Self::InvalidMerkleTreeNodeIndex { label, source_file, depth, index }
     }
 
     pub fn invalid_stack_depth_on_return(
@@ -472,31 +415,6 @@ impl ExecutionError {
             err_code,
             err_msg,
         }
-    }
-
-    pub fn merkle_store_lookup_failed(
-        err: MerkleError,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Self {
-        let (label, source_file) = err_ctx.label_and_source_file();
-        Self::MerkleStoreLookupFailed { label, source_file, err }
-    }
-
-    /// Note: This error currently never occurs, since `MerkleStore::merge_roots()` never fails.
-    pub fn merkle_store_merge_failed(
-        err: MerkleError,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Self {
-        let (label, source_file) = err_ctx.label_and_source_file();
-        Self::MerkleStoreMergeFailed { label, source_file, err }
-    }
-
-    pub fn merkle_store_update_failed(
-        err: MerkleError,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Self {
-        let (label, source_file) = err_ctx.label_and_source_file();
-        Self::MerkleStoreUpdateFailed { label, source_file, err }
     }
 
     pub fn no_mast_forest_with_procedure(