+++ processor/src/fast/io_ops.rs
@@ -1,5 +1,5 @@
 use super::{DOUBLE_WORD_SIZE, ExecutionError, FastProcessor, Felt, WORD_SIZE_FELT};
-use crate::{AdviceProvider, ErrorContext, Host, ProcessState};
+use crate::{AdviceProvider, ErrorContext, Host};
 
 impl FastProcessor {
     /// Analogous to `Process::op_push`.
@@ -10,9 +10,9 @@ impl FastProcessor {
 
     /// Analogous to `Process::op_advpop`.
     pub fn op_advpop(&mut self, op_idx: usize, host: &mut impl Host) -> Result<(), ExecutionError> {
-        let value = host
-            .advice_provider_mut()
-            .pop_stack(ProcessState::new_fast(self, op_idx), &ErrorContext::default())?;
+        let value = host.advice_provider_mut().pop_stack().map_err(|err| {
+            ExecutionError::advice_error_at_clk(err, self.clk + op_idx, &ErrorContext::default())
+        })?;
         self.increment_stack_size();
         self.stack_write(0, value);
         Ok(())
@@ -24,9 +24,9 @@ impl FastProcessor {
         op_idx: usize,
         host: &mut impl Host,
     ) -> Result<(), ExecutionError> {
-        let word = host
-            .advice_provider_mut()
-            .pop_stack_word(ProcessState::new_fast(self, op_idx), &ErrorContext::default())?;
+        let word = host.advice_provider_mut().pop_stack_word().map_err(|err| {
+            ExecutionError::advice_error_at_clk(err, self.clk + op_idx, &ErrorContext::default())
+        })?;
         self.stack_write_word(0, &word);
 
         Ok(())
@@ -109,9 +109,9 @@ impl FastProcessor {
         let addr_second_word = addr_first_word + WORD_SIZE_FELT;
 
         // pop two words from the advice stack
-        let words = host
-            .advice_provider_mut()
-            .pop_stack_dword(ProcessState::new_fast(self, op_idx), &ErrorContext::default())?;
+        let words = host.advice_provider_mut().pop_stack_dword().map_err(|err| {
+            ExecutionError::advice_error_at_clk(err, self.clk + op_idx, &ErrorContext::default())
+        })?;
 
         // write the words to memory
         self.memory.write_word(self.ctx, addr_first_word, self.clk + op_idx, words[0])?;