+++ processor/src/host/advice/mod.rs
@@ -1,14 +1,15 @@
 use alloc::vec::Vec;
 
-use vm_core::{Felt, crypto::merkle::MerklePath, mast::MastNodeExt};
-
-use crate::{ErrorContext, ExecutionError, ProcessState, Word};
+use vm_core::{Felt, Word, crypto::merkle::MerklePath};
 
 mod inputs;
 pub use inputs::AdviceInputs;
 
 mod providers;
 pub use providers::{MemAdviceProvider, RecAdviceProvider};
+
+mod errors;
+pub use errors::AdviceProviderError;
 
 mod source;
 pub use source::AdviceSource;
@@ -31,7 +32,7 @@ pub use source::AdviceSource;
 /// 3. Merkle store, which contains structured data reducible to Merkle paths. The VM can request
 ///    Merkle paths from the store, as well as mutate it by updating or merging nodes contained in
 ///    the store.
-pub trait AdviceProvider: Sized {
+pub trait AdviceProvider {
     // REQUIRED METHODS
     // --------------------------------------------------------------------------------------------
 
@@ -42,11 +43,7 @@ pub trait AdviceProvider: Sized {
     ///
     /// # Errors
     /// Returns an error if the advice stack is empty.
-    fn pop_stack(
-        &mut self,
-        process: ProcessState,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<Felt, ExecutionError>;
+    fn pop_stack(&mut self) -> Result<Felt, AdviceProviderError>;
 
     /// Pops a word (4 elements) from the advice stack and returns it.
     ///
@@ -55,11 +52,7 @@ pub trait AdviceProvider: Sized {
     ///
     /// # Errors
     /// Returns an error if the advice stack does not contain a full word.
-    fn pop_stack_word(
-        &mut self,
-        process: ProcessState,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<Word, ExecutionError>;
+    fn pop_stack_word(&mut self) -> Result<Word, AdviceProviderError>;
 
     /// Pops a double word (8 elements) from the advice stack and returns them.
     ///
@@ -69,21 +62,13 @@ pub trait AdviceProvider: Sized {
     ///
     /// # Errors
     /// Returns an error if the advice stack does not contain two words.
-    fn pop_stack_dword(
-        &mut self,
-        process: ProcessState,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<[Word; 2], ExecutionError>;
+    fn pop_stack_dword(&mut self) -> Result<[Word; 2], AdviceProviderError>;
 
     /// Pushes the value(s) specified by the source onto the advice stack.
     ///
     /// # Errors
     /// Returns an error if the value specified by the advice source cannot be obtained.
-    fn push_stack(
-        &mut self,
-        source: AdviceSource,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<(), ExecutionError>;
+    fn push_stack(&mut self, source: AdviceSource) -> Result<(), AdviceProviderError>;
 
     /// Returns a slice of length `length` from the top of the advice stack.
     /// If length = 0 returns the whole advice stack.
@@ -120,8 +105,7 @@ pub trait AdviceProvider: Sized {
         root: Word,
         depth: &Felt,
         index: &Felt,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<Word, ExecutionError>;
+    ) -> Result<Word, AdviceProviderError>;
 
     /// Returns a path to a node at the specified depth and index in a Merkle tree with the
     /// specified root.
@@ -137,8 +121,7 @@ pub trait AdviceProvider: Sized {
         root: Word,
         depth: &Felt,
         index: &Felt,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<MerklePath, ExecutionError>;
+    ) -> Result<MerklePath, AdviceProviderError>;
 
     /// Reconstructs a path from the root until a leaf or empty node and returns its depth.
     ///
@@ -153,8 +136,7 @@ pub trait AdviceProvider: Sized {
         root: Word,
         tree_depth: &Felt,
         index: &Felt,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<u8, ExecutionError>;
+    ) -> Result<u8, AdviceProviderError>;
 
     /// Updates a node at the specified depth and index in a Merkle tree with the specified root;
     /// returns the Merkle path from the updated node to the new root, together with the new root.
@@ -175,8 +157,7 @@ pub trait AdviceProvider: Sized {
         depth: &Felt,
         index: &Felt,
         value: Word,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<(MerklePath, Word), ExecutionError>;
+    ) -> Result<(MerklePath, Word), AdviceProviderError>;
 
     /// Creates a new Merkle tree in the advice provider by combining Merkle trees with the
     /// specified roots. The root of the new tree is defined as `hash(left_root, right_root)`.
@@ -186,10 +167,5 @@ pub trait AdviceProvider: Sized {
     ///
     /// It is not checked whether a Merkle tree for either of the specified roots can be found in
     /// this advice provider.
-    fn merge_roots(
-        &mut self,
-        lhs: Word,
-        rhs: Word,
-        err_ctx: &ErrorContext<'_, impl MastNodeExt>,
-    ) -> Result<Word, ExecutionError>;
+    fn merge_roots(&mut self, lhs: Word, rhs: Word) -> Result<Word, AdviceProviderError>;
 }