+++ processor/src/lib.rs
@@ -54,8 +54,14 @@ use range::RangeChecker;
 
 mod host;
 pub use host::{
-    DefaultHost, Host, MastForestStore, MemMastForestStore,
-    advice::{AdviceInputs, AdviceProvider, AdviceSource, MemAdviceProvider, RecAdviceProvider},
+    Host, HostLibrary, MastForestStore, MemMastForestStore,
+    advice::{
+        AdviceInputs, AdviceProvider, AdviceProviderError, AdviceSource, MemAdviceProvider,
+        RecAdviceProvider,
+    },
+    default,
+    default::DefaultHost,
+    handlers,
 };
 
 mod chiplets;
@@ -838,13 +844,11 @@ impl<'a> ProcessState<'a> {
     /// # Errors
     /// - If the address is not word aligned.
     #[inline(always)]
-    pub fn get_mem_word(&self, ctx: ContextId, addr: u32) -> Result<Option<Word>, ExecutionError> {
+    pub fn get_mem_word(&self, ctx: ContextId, addr: u32) -> Result<Option<Word>, MemoryError> {
         match self {
-            ProcessState::Slow(state) => {
-                state.chiplets.memory.get_word(ctx, addr).map_err(ExecutionError::MemoryError)
-            },
+            ProcessState::Slow(state) => state.chiplets.memory.get_word(ctx, addr),
             ProcessState::Fast(state) => {
-                Ok(state.processor.memory.read_word_impl(ctx, addr, None)?.copied())
+                state.processor.memory.read_word_impl(ctx, addr, None).map(|w| w.copied())
             },
         }
     }